Fix arm_darwin build: explicit cast of `cnt` to `long double` in `MarkovModel`
The `arm_darwin` build failed to compile `src/Client/MarkovModel.cpp` with
`-Werror,-Wimplicit-int-float-conversion`:

    error: implicit conversion from 'const size_t' (aka 'const unsigned long')
    to 'long double' may lose precision

In `NGram::UnigramsWithStats::getWeightedWordCount`, `cnt` (a `size_t`) was
implicitly converted to `long double` in the multiplication, while only
`last_timestamp` was cast explicitly. On x86-64 `long double` is 80-bit and
can represent a 64-bit `size_t` exactly (no warning), but on ARM Darwin
`long double` is 64-bit (`double`), so the conversion is flagged. Add an
explicit `static_cast<long double>` around `cnt`, mirroring the identical
pattern already used in `KneserNey::WordWithCount::getWeightedScore`.

CI report: https://s3.amazonaws.com/clickhouse-test-reports/json.html?PR=109244&sha=4661afac6526f9cf969ae93ded0d65195bb7aed3&name_0=PR&name_1=Build%20%28arm_darwin%29

+++ src/Client/MarkovModel.cpp
@@ -84,7 +84,7 @@ long double NGram::UnigramsWithStats::getWeightedWordCount(const std::string & w
 {
     if (unigram_map.contains(word))
     {
-        return unigram_map.at(word).cnt * std::sqrt(static_cast<long double>(unigram_map.at(word).last_timestamp));
+        return static_cast<long double>(unigram_map.at(word).cnt) * std::sqrt(static_cast<long double>(unigram_map.at(word).last_timestamp));
     }
     return 0.0L;
 }
